issue #1 - snapshot: move code around, get ready to wire in new georeference code in to tools
