+++ georeference/assign_test.go
